Fixes to review item, some naming conventions
