+++ .gitignore
@@ -1,3 +1,5 @@
 # IDE preferences
 .vscode/
 .idea/
+
+/node_modules/
