+++ frontend/package.json
@@ -18,6 +18,7 @@
     "bootstrap": "^5.3.3",
     "mdb-react-ui-kit": "^8.0.0",
     "react": "^18.3.1",
+    "react-bootstrap": "^2.10.2",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.23.0",
     "react-scripts": "5.0.1",
